Module epic_wallet_libwallet::api_impl::owner

Generic implementation of owner API functions

Functions

accounts

List of accounts

cancel_tx

cancel tx

check_ttl

Check TTL

create_account_path

new account path

finalize_tx

Finalize slate

get_public_proof_address

Retrieve the payment proof address for the current parent key at the given index set active account

get_stored_tx

get stored tx

init_send_tx

Initiate tx as sender

issue_invoice_tx

Initiate a transaction as the recipient (invoicing)

node_height

node height

post_tx

Posts a transaction to the chain take a client impl instead of wallet so as not to have to lock the wallet

process_invoice_tx

Receive an invoice tx, essentially adding inputs to whatever output was specified

retrieve_outputs

retrieve outputs

retrieve_summary_info

Retrieve summary info

retrieve_txs

Retrieve txs

scan

check repair Accepts a wallet inst instead of a raw wallet so it can lock as little as possible

set_active_account

set active account

tx_lock_outputs

Lock sender outputs

update_wallet_state

Experimental, wrap the entire definition of how a wallet’s state is updated

verify_slate_messages

verify slate messages